updated priceBucket.liquidity to decimal

+++ OkonkwoOandaV20/OkonkwoOandaV20/TradeLibrary/Pricing/PriceBucket.cs
@@ -13,6 +13,6 @@
       /// <summary>
       /// The amount of liquidity offered by the PriceBucket
       /// </summary>
-      public int liquidity { get; set; }
+      public decimal liquidity { get; set; }
    }
 }
